[fix] No.106 Chapter18
* プロパティの宣言順番
* クラスのファイル分け
